Collapse app/session/ folder; move files into systems/ (refs #1200)
Per #1200, the app/session/ folder was a 'system in costume': it bundled
init/teardown systems that operate on entt::registry as free functions —
the canonical shape of an ECS system. Same pattern previously closed for
app/input/ (#1216) and app/audio/ (#1217).

Moves (git-tracked renames):
  app/session/play_session.{h,cpp}        -> app/systems/play_session.{h,cpp}
  app/session/test_player_session.{h,cpp} -> app/systems/test_player_session.{h,cpp}

Include updates:
  app/systems/play_session.cpp: drop ../systems/ prefix from sibling
    all_systems.h, high_score_system.h, game_phase_transition.h includes.
  app/systems/test_player_session.cpp: drop ../systems/ prefix from
    sibling session_logger_system.h.
  app/systems/game_state_system.cpp: ../session/play_session.h ->
    play_session.h.
  app/game_loop.cpp: session/test_player_session.h ->
    systems/test_player_session.h.
  tests/test_high_score_integration.cpp: session/play_session.h ->
    systems/play_session.h.
  tests/test_test_player_system.cpp: session/test_player_session.h ->
    systems/test_player_session.h.
  tests/test_phase_transition_canonical.cpp: canonical enter_phase
    allowlist + INFO message updated to app/systems/play_session.cpp.

Build:
  CMakeLists.txt: SESSION_SOURCES glob deleted; relocated .cpp files
  now picked up by the existing SYSTEM_SOURCES glob. app/session/
  directory deleted.

Validation:
  - cmake --build build (unity): clean, zero warnings.
  - cmake --build build-nonunity (CMAKE_UNITY_BUILD=OFF): clean.
  - ./build/shapeshifter_tests: 901 Catch2 cases / 2957 assertions PASS.
  - ctest -j: 912/913 PASS (1 smoke test skipped by default).
  - [phase_transition] tag: 4/4 cases PASS (allowlist test verified).
