+++ app/systems/game_state_system.cpp
@@ -1,7 +1,7 @@
 #include "all_systems.h"
 #include "game_phase_transition.h"
 #include "game_state_system.h"
-#include "../session/play_session.h"
+#include "play_session.h"
 #include "../components/game_state.h"
 #include "../components/input.h"
 #include "../components/obstacle.h"
