+++ app/systems/play_session.cpp
@@ -8,15 +8,15 @@
 #include "../components/song_state.h"
 #include "../components/music.h"
 #include "../components/high_score.h"
-#include "../systems/high_score_system.h"
+#include "high_score_system.h"
 #include "../entities/settings.h"
 #include "../util/rhythm_math.h"
 #include "../entities/camera_entity.h"
 #include "../entities/energy_bar_entity.h"
 #include "../entities/player_entity.h"
 #include "../content/level_content_config.h"
-#include "../systems/all_systems.h"
-#include "../systems/game_phase_transition.h"
+#include "all_systems.h"
+#include "game_phase_transition.h"
 #include "../constants.h"
 #include <filesystem>
 #include <raylib.h>
